30/07/2021, Confirmacao de delecao de uma lista, atraves de dialog confirmation!

+++ virtusApp/src/app/admin/configuracao/ciclos/ciclo-list/ciclo-list.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
-import {CicloService} from '../ciclo.service';
+import { CicloService } from '../ciclo.service';
 import { Ciclo } from '../ciclo';
 import { Router } from '@angular/router';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { UpdateCicloComponent } from '../update-ciclo/update-ciclo.component';
+import { DialogService } from 'src/app/services/dialog.service';
+import { NotificationService } from 'src/app/services/notification.service';
 
 @Component({
   selector: 'app-ciclo-list',
@@ -16,7 +18,10 @@ export class CicloListComponent implements OnInit {
 	ciclos!: Ciclo[] 
 	
   constructor(private cicloService: CicloService,  
-	private router:Router, public matDialog: MatDialog) { }
+	private router:Router,
+	public matDialog: MatDialog,
+	private dialogService:DialogService,
+	private notificationService:NotificationService) { }
 
   ngOnInit(): void {
 	this.getCiclos();
@@ -25,6 +30,7 @@ export class CicloListComponent implements OnInit {
    private getCiclos(){
 		this.cicloService.getCiclosList().subscribe(data => {
 			this.ciclos = data;
+			
 			});
 	}
 
@@ -39,10 +45,18 @@ export class CicloListComponent implements OnInit {
 	}
 
 	deleteCiclo(id: number){
-		this.cicloService.deleteCiclo(id).subscribe(data => {
-		console.log('Deletou isso:',data);
-		this.getCiclos();
-		})
+		this.dialogService.openConfirmDialog('Tem certeza que deseja apagar o Registro??')
+		.afterClosed().subscribe(res =>{
+			if (res){
+				console.log('!');
+				this.cicloService.deleteCiclo(id).subscribe(data => {
+					console.log('Deletou isso:',data);
+					this.getCiclos();
+					this.notificationService.warn('Ciclo deletado com sucesso!')
+				})
+			}
+		});
+		
 	}
 	openModalUpdate(id:number) {
 		const dialogConfig = new MatDialogConfig();
@@ -55,14 +69,5 @@ export class CicloListComponent implements OnInit {
 		const modalDialog = this.matDialog.open(UpdateCicloComponent, dialogConfig);
 	  }
 
-	  openModalDelete(id:number) {
-		const dialogConfig = new MatDialogConfig();
-		// The user can't close the dialog by clicking outside its body
-		dialogConfig.disableClose = false;
-		dialogConfig.id = "modal-component", id;
-		dialogConfig.height = "350px";
-		dialogConfig.width = "500px";
-		// https://material.angular.io/components/dialog/overview
-		const modalDialog = this.matDialog.open(deleteCiclo(id));
-	  }
+
 }
